requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fsspec==2025.7.0
gradio==5.38.1
gradio_client==1.11.0
groovy==0.1.2
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.1
openai==1.97.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.12.4
safehttpx==0.1.6
scikit-learn==1.7.1
scipy==1.16.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
